split line routes to separate file
