+++ server/data/lines.ts
@@ -1,655 +1,140 @@
-import {
-  BranchingLineRoute,
-  Line,
-  LoopLineRoute,
-  SimpleLineRoute,
-  WerribeeLineRoute,
-} from "./line";
+import { Line } from "./line";
 import { LineCollection } from "./line-collection";
 import * as line from "./line-ids";
-import * as station from "./station-ids";
+import * as route from "./line-routes";
 
 const data = [
   new Line({
     id: line.SANDRINGHAM,
     name: "Sandringham",
     ptvIds: [12],
-    route: new SimpleLineRoute([
-      station.FLINDERS_STREET,
-      station.RICHMOND,
-      station.SOUTH_YARRA,
-      station.PRAHRAN,
-      station.WINDSOR,
-      station.BALACLAVA,
-      station.RIPPONLEA,
-      station.ELSTERNWICK,
-      station.GARDENVALE,
-      station.NORTH_BRIGHTON,
-      station.MIDDLE_BRIGHTON,
-      station.BRIGHTON_BEACH,
-      station.HAMPTON,
-      station.SANDRINGHAM,
-    ]),
+    route: route.SANDRINGHAM,
   }),
   new Line({
     id: line.FRANKSTON,
     name: "Frankston",
     ptvIds: [6],
-    route: new SimpleLineRoute([
-      station.FLINDERS_STREET,
-      station.RICHMOND,
-      station.SOUTH_YARRA,
-      station.HAWKSBURN,
-      station.TOORAK,
-      station.ARMADALE,
-      station.MALVERN,
-      station.CAULFIELD,
-      station.GLEN_HUNTLY,
-      station.ORMOND,
-      station.MCKINNON,
-      station.BENTLEIGH,
-      station.PATTERSON,
-      station.MOORABBIN,
-      station.HIGHETT,
-      station.SOUTHLAND,
-      station.CHELTENHAM,
-      station.MENTONE,
-      station.PARKDALE,
-      station.MORDIALLOC,
-      station.ASPENDALE,
-      station.EDITHVALE,
-      station.CHELSEA,
-      station.BONBEACH,
-      station.CARRUM,
-      station.SEAFORD,
-      station.KANANOOK,
-      station.FRANKSTON,
-    ]),
+    route: route.FRANKSTON,
   }),
   new Line({
     id: line.STONY_POINT,
     name: "Stony Point",
     ptvIds: [13],
-    route: new SimpleLineRoute([
-      station.FRANKSTON,
-      station.LEAWARRA,
-      station.BAXTER,
-      station.SOMERVILLE,
-      station.TYABB,
-      station.HASTINGS,
-      station.BITTERN,
-      station.MORRADOO,
-      station.CRIB_POINT,
-      station.STONY_POINT,
-    ]),
+    route: route.STONY_POINT,
   }),
   new Line({
     id: line.CRANBOURNE,
     name: "Cranbourne",
     ptvIds: [4],
-    route: new LoopLineRoute([
-      station.RICHMOND,
-      station.SOUTH_YARRA,
-      station.HAWKSBURN,
-      station.TOORAK,
-      station.ARMADALE,
-      station.MALVERN,
-      station.CAULFIELD,
-      station.CARNEGIE,
-      station.MURRUMBEENA,
-      station.HUGHESDALE,
-      station.OAKLEIGH,
-      station.HUNTINGDALE,
-      station.CLAYTON,
-      station.WESTALL,
-      station.SPRINGVALE,
-      station.SANDOWN_PARK,
-      station.NOBLE_PARK,
-      station.YARRAMAN,
-      station.DANDENONG,
-      station.LYNBROOK,
-      station.MERINDA_PARK,
-      station.CRANBOURNE,
-    ]),
+    route: route.CRANBOURNE,
   }),
   new Line({
     id: line.PAKENHAM,
     name: "Pakenham",
     ptvIds: [11],
-    route: new LoopLineRoute([
-      station.RICHMOND,
-      station.SOUTH_YARRA,
-      station.HAWKSBURN,
-      station.TOORAK,
-      station.ARMADALE,
-      station.MALVERN,
-      station.CAULFIELD,
-      station.CARNEGIE,
-      station.MURRUMBEENA,
-      station.HUGHESDALE,
-      station.OAKLEIGH,
-      station.HUNTINGDALE,
-      station.CLAYTON,
-      station.WESTALL,
-      station.SPRINGVALE,
-      station.SANDOWN_PARK,
-      station.NOBLE_PARK,
-      station.YARRAMAN,
-      station.DANDENONG,
-      station.HALLAM,
-      station.NARRE_WARREN,
-      station.BERWICK,
-      station.BEACONSFIELD,
-      station.OFFICER,
-      station.CARDINIA_ROAD,
-      station.PAKENHAM,
-      station.EAST_PAKENHAM,
-    ]),
+    route: route.PAKENHAM,
   }),
   new Line({
     id: line.GIPPSLAND,
     name: "Gippsland",
     ptvIds: [1823, 1824],
-    route: new SimpleLineRoute([
-      station.SOUTHERN_CROSS,
-      station.FLINDERS_STREET,
-      station.RICHMOND,
-      station.SOUTH_YARRA,
-      station.HAWKSBURN,
-      station.TOORAK,
-      station.ARMADALE,
-      station.MALVERN,
-      station.CAULFIELD,
-      station.CARNEGIE,
-      station.MURRUMBEENA,
-      station.HUGHESDALE,
-      station.OAKLEIGH,
-      station.HUNTINGDALE,
-      station.CLAYTON,
-      station.WESTALL,
-      station.SPRINGVALE,
-      station.SANDOWN_PARK,
-      station.NOBLE_PARK,
-      station.YARRAMAN,
-      station.DANDENONG,
-      station.HALLAM,
-      station.NARRE_WARREN,
-      station.BERWICK,
-      station.BEACONSFIELD,
-      station.OFFICER,
-      station.CARDINIA_ROAD,
-      station.PAKENHAM,
-      station.EAST_PAKENHAM,
-      station.NAR_NAR_GOON,
-      station.TYNONG,
-      station.GARFIELD,
-      station.BUNYIP,
-      station.LONGWARRY,
-      station.DROUIN,
-      station.WARRAGUL,
-      station.YARRAGON,
-      station.TRAFALGAR,
-      station.MOE,
-      station.MORWELL,
-      station.TRARALGON,
-      station.ROSEDALE,
-      station.SALE,
-      station.STRATFORD,
-      station.BAIRNSDALE,
-    ]),
+    route: route.SANDRINGHAM,
   }),
   new Line({
     id: line.GLEN_WAVERLEY,
     name: "Glen Waverley",
     ptvIds: [7],
-    route: new LoopLineRoute([
-      station.RICHMOND,
-      station.EAST_RICHMOND,
-      station.BURNLEY,
-      station.HEYINGTON,
-      station.KOOYONG,
-      station.TOORONGA,
-      station.GARDINER,
-      station.GLEN_IRIS,
-      station.DARLING,
-      station.EAST_MALVERN,
-      station.HOLMESGLEN,
-      station.JORDANVILLE,
-      station.MOUNT_WAVERLEY,
-      station.SYNDAL,
-      station.GLEN_WAVERLEY,
-    ]),
+    route: route.GLEN_WAVERLEY,
   }),
   new Line({
     id: line.ALAMEIN,
     name: "Alamein",
     ptvIds: [1],
-    route: new LoopLineRoute([
-      station.RICHMOND,
-      station.EAST_RICHMOND,
-      station.BURNLEY,
-      station.HAWTHORN,
-      station.GLENFERRIE,
-      station.AUBURN,
-      station.CAMBERWELL,
-      station.RIVERSDALE,
-      station.WILLISON,
-      station.HARTWELL,
-      station.BURWOOD,
-      station.ASHBURTON,
-      station.ALAMEIN,
-    ]),
+    route: route.ALAMEIN,
   }),
   new Line({
     id: line.BELGRAVE,
     name: "Belgrave",
     ptvIds: [2],
-    route: new LoopLineRoute([
-      station.RICHMOND,
-      station.EAST_RICHMOND,
-      station.BURNLEY,
-      station.HAWTHORN,
-      station.GLENFERRIE,
-      station.AUBURN,
-      station.CAMBERWELL,
-      station.EAST_CAMBERWELL,
-      station.CANTERBURY,
-      station.CHATHAM,
-      station.UNION,
-      station.BOX_HILL,
-      station.LABURNUM,
-      station.BLACKBURN,
-      station.NUNAWADING,
-      station.MITCHAM,
-      station.HEATHERDALE,
-      station.RINGWOOD,
-      station.HEATHMONT,
-      station.BAYSWATER,
-      station.BORONIA,
-      station.FERNTREE_GULLY,
-      station.UPPER_FERNTREE_GULLY,
-      station.UPWEY,
-      station.TECOMA,
-      station.BELGRAVE,
-    ]),
+    route: route.BELGRAVE,
   }),
   new Line({
     id: line.LILYDALE,
     name: "Lilydale",
     ptvIds: [9],
-    route: new LoopLineRoute([
-      station.RICHMOND,
-      station.EAST_RICHMOND,
-      station.BURNLEY,
-      station.HAWTHORN,
-      station.GLENFERRIE,
-      station.AUBURN,
-      station.CAMBERWELL,
-      station.EAST_CAMBERWELL,
-      station.CANTERBURY,
-      station.CHATHAM,
-      station.UNION,
-      station.BOX_HILL,
-      station.LABURNUM,
-      station.BLACKBURN,
-      station.NUNAWADING,
-      station.MITCHAM,
-      station.HEATHERDALE,
-      station.RINGWOOD,
-      station.RINGWOOD_EAST,
-      station.CROYDON,
-      station.MOOROOLBARK,
-      station.LILYDALE,
-    ]),
+    route: route.LILYDALE,
   }),
   new Line({
     id: line.HURSTBRIDGE,
     name: "Hurstbridge",
     ptvIds: [8],
-    route: new LoopLineRoute([
-      station.JOLIMONT,
-      station.WEST_RICHMOND,
-      station.NORTH_RICHMOND,
-      station.COLLINGWOOD,
-      station.VICTORIA_PARK,
-      station.CLIFTON_HILL,
-      station.WESTGARTH,
-      station.DENNIS,
-      station.FAIRFIELD,
-      station.ALPHINGTON,
-      station.DAREBIN,
-      station.IVANHOE,
-      station.EAGLEMONT,
-      station.HEIDELBERG,
-      station.ROSANNA,
-      station.MACLEOD,
-      station.WATSONIA,
-      station.GREENSBOROUGH,
-      station.MONTMORENCY,
-      station.ELTHAM,
-      station.DIAMOND_CREEK,
-      station.WATTLE_GLEN,
-      station.HURSTBRIDGE,
-    ]),
+    route: route.HURSTBRIDGE,
   }),
   new Line({
     id: line.MERNDA,
     name: "Mernda",
     ptvIds: [5],
-    route: new LoopLineRoute([
-      station.JOLIMONT,
-      station.WEST_RICHMOND,
-      station.NORTH_RICHMOND,
-      station.COLLINGWOOD,
-      station.VICTORIA_PARK,
-      station.CLIFTON_HILL,
-      station.RUSHALL,
-      station.MERRI,
-      station.NORTHCOTE,
-      station.CROXTON,
-      station.THORNBURY,
-      station.BELL,
-      station.PRESTON,
-      station.REGENT,
-      station.RESERVOIR,
-      station.RUTHVEN,
-      station.KEON_PARK,
-      station.THOMASTOWN,
-      station.LALOR,
-      station.EPPING,
-      station.SOUTH_MORANG,
-      station.MIDDLE_GORGE,
-      station.HAWKSTOWE,
-      station.MERNDA,
-    ]),
+    route: route.MERNDA,
   }),
   new Line({
     id: line.UPFIELD,
     name: "Upfield",
     ptvIds: [15],
-    route: new LoopLineRoute([
-      station.NORTH_MELBOURNE,
-      station.MACAULAY,
-      station.FLEMINGTON_BRIDGE,
-      station.ROYAL_PARK,
-      station.JEWELL,
-      station.BRUNSWICK,
-      station.ANSTEY,
-      station.MORELAND,
-      station.COBURG,
-      station.BATMAN,
-      station.MERLYNSTON,
-      station.FAWKNER,
-      station.GOWRIE,
-      station.UPFIELD,
-    ]),
+    route: route.UPFIELD,
   }),
   new Line({
     id: line.CRAIGIEBURN,
     name: "Craigieburn",
     ptvIds: [3],
-    route: new LoopLineRoute([
-      station.NORTH_MELBOURNE,
-      station.KENSINGTON,
-      station.NEWMARKET,
-      station.ASCOT_VALE,
-      station.MOONEE_PONDS,
-      station.ESSENDON,
-      station.GLENBERVIE,
-      station.STRATHMORE,
-      station.PASCOE_VALE,
-      station.OAK_PARK,
-      station.GLENROY,
-      station.JACANA,
-      station.BROADMEADOWS,
-      station.COOLAROO,
-      station.ROXBURGH_PARK,
-      station.CRAIGIEBURN,
-    ]),
+    route: route.CRAIGIEBURN,
   }),
   new Line({
     id: line.SEYMOUR,
     name: "Seymour",
     ptvIds: [1706, 1710, 1908],
-    route: new BranchingLineRoute(
-      [
-        station.SOUTHERN_CROSS,
-        station.NORTH_MELBOURNE,
-        station.KENSINGTON,
-        station.NEWMARKET,
-        station.ASCOT_VALE,
-        station.MOONEE_PONDS,
-        station.ESSENDON,
-        station.GLENBERVIE,
-        station.STRATHMORE,
-        station.PASCOE_VALE,
-        station.OAK_PARK,
-        station.GLENROY,
-        station.JACANA,
-        station.BROADMEADOWS,
-        station.COOLAROO,
-        station.ROXBURGH_PARK,
-        station.CRAIGIEBURN,
-        station.DONNYBROOK,
-        station.WALLAN,
-        station.HEATHCOTE_JUNCTION,
-        station.WANDONG,
-        station.KILMORE_EAST,
-        station.BROADFORD,
-        station.TALLAROOK,
-        station.SEYMOUR,
-      ],
-      [
-        station.AVENEL,
-        station.EUROA,
-        station.VIOLET_TOWN,
-        station.BENALLA,
-        station.WANGARATTA,
-        station.SPRINGHURST,
-        station.CHILTERN,
-        station.WODONGA,
-        station.ALBURY,
-      ],
-      [
-        station.NAGAMBIE,
-        station.MURCHISON_EAST,
-        station.MOOROOPNA,
-        station.SHEPPARTON,
-      ],
-    ),
+    route: route.SEYMOUR,
   }),
   new Line({
     id: line.FLEMINGTON_RACECOURSE,
     name: "Flemington Racecourse",
     ptvIds: [1482],
-    route: new SimpleLineRoute([
-      station.FLINDERS_STREET,
-      station.SOUTHERN_CROSS,
-      station.NORTH_MELBOURNE,
-      station.KENSINGTON,
-      station.NEWMARKET,
-      station.SHOWGROUNDS,
-      station.FLEMINGTON_RACECOURSE,
-    ]),
+    route: route.FLEMINGTON_RACECOURSE,
   }),
   new Line({
     id: line.SUNBURY,
     name: "Sunbury",
     ptvIds: [14],
-    route: new LoopLineRoute([
-      station.NORTH_MELBOURNE,
-      station.SOUTH_KENSINGTON,
-      station.FOOTSCRAY,
-      station.MIDDLE_FOOTSCRAY,
-      station.WEST_FOOTSCRAY,
-      station.TOTTENHAM,
-      station.SUNSHINE,
-      station.ALBION,
-      station.GINIFER,
-      station.ST_ALBANS,
-      station.KEILOR_PLAINS,
-      station.WATERGARDENS,
-      station.DIGGERS_REST,
-      station.SUNBURY,
-    ]),
+    route: route.SUNBURY,
   }),
   new Line({
     id: line.BENDIGO,
     name: "Bendigo",
     ptvIds: [1740, 1848, 1849],
-    route: new BranchingLineRoute(
-      [
-        station.SOUTHERN_CROSS,
-        station.NORTH_MELBOURNE,
-        station.SOUTH_KENSINGTON,
-        station.FOOTSCRAY,
-        station.MIDDLE_FOOTSCRAY,
-        station.WEST_FOOTSCRAY,
-        station.TOTTENHAM,
-        station.SUNSHINE,
-        station.ALBION,
-        station.GINIFER,
-        station.ST_ALBANS,
-        station.KEILOR_PLAINS,
-        station.WATERGARDENS,
-        station.DIGGERS_REST,
-        station.SUNBURY,
-        station.CLARKEFIELD,
-        station.RIDDELLS_CREEK,
-        station.GISBORNE,
-        station.MACEDON,
-        station.WOODEND,
-        station.KYNETON,
-        station.MALMSBURY,
-        station.CASTLEMAINE,
-        station.KANGAROO_FLAT,
-        station.BENDIGO,
-      ],
-      [
-        station.EPSOM,
-        station.HUNTLY,
-        station.GOORNONG,
-        station.ELMORE,
-        station.ROCHESTER,
-        station.ECHUCA,
-      ],
-      [
-        station.EAGLEHAWK,
-        station.RAYWOOD,
-        station.DINGEE,
-        station.PYRAMID,
-        station.KERANG,
-        station.SWAN_HILL,
-      ],
-    ),
+    route: route.BENDIGO,
   }),
   new Line({
     id: line.BALLARAT,
     name: "Ballarat",
     ptvIds: [1728, 1837, 4871],
-    route: new BranchingLineRoute(
-      [
-        station.SOUTHERN_CROSS,
-        station.NORTH_MELBOURNE,
-        station.SOUTH_KENSINGTON,
-        station.FOOTSCRAY,
-        station.MIDDLE_FOOTSCRAY,
-        station.WEST_FOOTSCRAY,
-        station.TOTTENHAM,
-        station.SUNSHINE,
-        station.ARDEER,
-        station.DEER_PARK,
-        station.CAROLINE_SPRINGS,
-        station.ROCKBANK,
-        station.COBBLEBANK,
-        station.MELTON,
-        station.BACCHUS_MARSH,
-        station.BALLAN,
-        station.BALLARAT,
-      ],
-      [station.CRESWICK, station.CLUNES, station.TALBOT, station.MARYBOROUGH],
-      [station.WENDOUREE, station.BEAUFORT, station.ARARAT],
-    ),
+    route: route.BALLARAT,
   }),
   new Line({
     id: line.GEELONG,
     name: "Geelong",
     ptvIds: [1745, 1853],
-    route: new SimpleLineRoute([
-      station.SOUTHERN_CROSS,
-      station.NORTH_MELBOURNE,
-      station.SOUTH_KENSINGTON,
-      station.FOOTSCRAY,
-      station.MIDDLE_FOOTSCRAY,
-      station.WEST_FOOTSCRAY,
-      station.TOTTENHAM,
-      station.SUNSHINE,
-      station.ARDEER,
-      station.DEER_PARK,
-      station.TARNEIT,
-      station.WYNDHAM_VALE,
-      station.LITTLE_RIVER,
-      station.LARA,
-      station.CORIO,
-      station.NORTH_SHORE,
-      station.NORTH_GEELONG,
-      station.GEELONG,
-      station.SOUTH_GEELONG,
-      station.MARSHALL,
-      station.WAURN_PONDS,
-      station.WINCHELSEA,
-      station.BIRREGURRA,
-      station.COLAC,
-      station.CAMPERDOWN,
-      station.TERANG,
-      station.SHERWOOD_PARK,
-      station.WARRNAMBOOL,
-    ]),
+    route: route.GEELONG,
   }),
   new Line({
     id: line.WERRIBEE,
     name: "Werribee",
     ptvIds: [16],
-    route: new WerribeeLineRoute(
-      [
-        station.FLINDERS_STREET,
-        station.SOUTHERN_CROSS,
-        station.NORTH_MELBOURNE,
-        station.SOUTH_KENSINGTON,
-        station.FOOTSCRAY,
-        station.SEDDON,
-        station.YARRAVILLE,
-        station.SPOTSWOOD,
-        station.NEWPORT,
-      ],
-      [station.SEAHOLME, station.ALTONA, station.WESTONA],
-      [
-        station.LAVERTON,
-        station.AIRCRAFT,
-        station.WILLIAMS_LANDING,
-        station.HOPPERS_CROSSING,
-        station.WERRIBEE,
-      ],
-    ),
+    route: route.WERRIBEE,
   }),
   new Line({
     id: line.WILLIAMSTOWN,
     name: "Williamstown",
     ptvIds: [17],
-    route: new SimpleLineRoute([
-      station.FLINDERS_STREET,
-      station.SOUTHERN_CROSS,
-      station.NORTH_MELBOURNE,
-      station.SOUTH_KENSINGTON,
-      station.FOOTSCRAY,
-      station.SEDDON,
-      station.YARRAVILLE,
-      station.SPOTSWOOD,
-      station.NEWPORT,
-      station.NORTH_WILLIAMSTOWN,
-      station.WILLIAMSTOWN_BEACH,
-      station.WILLIAMSTOWN,
-    ]),
+    route: route.WILLIAMSTOWN,
   }),
 ];
